Object id is a long int
git-svn-id: file:///home/roberto/Git/svn@7085 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: 0e1681001c5d64aea15cd1e8943f8e45d89edfc3

+++ trunk/xmipp/libraries/data/image_base.cpp
@@ -66,7 +66,7 @@ int ImageBase::read2(const FileName &name, bool readdata, int select_img, bool a
 
 /** Read an image from metadata, filename is provided
 */
-int ImageBase::read(const FileName &name, const MetaData &md, int objId, bool readdata,
+int ImageBase::read(const FileName &name, const MetaData &md, long int objId, bool readdata,
                     int select_img, bool only_apply_shifts, bool mapData)
 {
     ImageFHandler* hFile = openFile(name);
